+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - use [ecow](https://github.com/typst/ecow) EcoString for the small strings that get cloned frequently
   - replace [strfmt](https://github.com/vitiral/strfmt) with [figura](https://github.com/saverioscagnoli/figura) for dynamic string formatting
   - improve log time display (rfc 2822 date + time in localtime)
+  - also log CLI arguments
 
 - 1.12.11 (Aug 1 2025 dheijl)
   - more small refactorings